filter packages to instrument (#1)
To enhance fuzzing performance, we can limit coverage instrumentation
to packages that we want the fuzzer to explore. This way, the fuzzer
will focus on exploring the packages of interest. Otherwise, libfuzzer
will try to maximize coverage in all packages. This PR limit coverage
instrumentation to the packages explicitly configured by the `-include`
flag (default is include all packages) and not excluded by the
`-preserve` flag (default is none)